UI and OS installation refactoring
Added a dedicated 'select' checkbox: OS can be selected without having to install it on the partition at all (if already done).
Replaces the indeterminate state of the 'install' checkbox.
Nothing is done for unselected OSes.

Merged the 'format' checkbox with the 'install' checkbox as the indeterminate state.
Used to format/erase (and install from ISO) partition, or leave it as-is.
If OS installation is not activated, ISO is not mounted.

Added a 'setup' checkbox to make OS setup optional.
Setup is still forced if OS installation is activated.

Added a 'bootloader' checkbox to make partition boot setup optional.
Bootloader is still forced if OS installation is activated.

Some other code refactoring and cleanup.