Remember chosen device UUID, preferred over path
The path depends on the order in which devices are detected or
plugged, and thus may change between runs.
The UUID, when available, does not change except when re-creating
the partition table: this happens rarely, as re-partitioning is
often enough when needing to re-do all partitions.

Thus remember both path and UUID for selected device.
If UUID is found in available devices, use it. Otherwise fallbacks
to matching the device path as was done previously.

Handle legacy setting migration.

For this to actually be really useful, we should also remember the
partition number (and not only path) so that we can adapt to the
new device path if needed.